vector_selection_internal.h: Move implementations from header to the .cc
